refactor(config:install): use access(2) for bin-dir writability check
Replace the temp-file probe in isWritableDir with golang.org/x/sys/unix
Access on Unix, falling back to the probe file only on Windows where no
equivalent syscall is available. The temp file was created and deleted
on every config:install invocation across each allowlist entry; the
permission-bit check has the same semantics without the side effect.
